task-16-default-export: created module with default export of class Encyclopedia
